Update README.md
This innovative project utilizes the SX1278 model of LoRa for establishing long-range wireless communication between two Arduino UNO boards. The transmission Arduino is connected to a Raspberry Pi through serial communication, and the Raspberry Pi is integrated with a programmable logic controller (PLC) that sends specific commands. The Raspberry Pi transmits these commands to the transmission Arduino via serial communication, and the latter uses the LoRa SX1278 module to transmit these commands to the receiving Arduino.

The transmitted commands include instructions such as activating and stopping a solenoid valve, thereby contributing to the optimization of irrigation resources. The open-source nature of the system provides a flexible and reliable approach to environmental monitoring and smart agriculture, addressing current challenges related to food security and efficient water management.